+++ src/data/SummaryTransform.ts
@@ -14,7 +14,8 @@ export interface SummaryChartMetricValue {
 }
 
 export interface SummaryChartData {
-  mapName: string;
+  displayName: string;
+  group?: string;
   totalAverage: number;
   metrics: Array<MetricEnum>;
   metricValues: SummaryChartMetricValue[];
@@ -59,7 +60,8 @@ export function buildSummaryChartData(
 
   if (!wholeUpdateAgg) {
     return {
-      mapName: result.fileName,
+      displayName: result.displayName,
+      group: result.group,
       metrics: metricValues.map(it => MetricRegistryInstance.getOrThrow(it.metricName)),
       metricValues,
       totalAverage: sumOfParts,
@@ -86,7 +88,8 @@ export function buildSummaryChartData(
   });
 
   return {
-    mapName: result.fileName,
+    displayName: result.displayName,
+    group: result.group,
     metrics: metricValues.map(it => MetricRegistryInstance.getOrThrow(it.metricName)),
     metricValues,
     totalAverage: wholeUpdateAverage,